rename Next -> Batching, add to some CI jobs
